+++ examples/bucket_policy/main.tf
@@ -16,7 +16,7 @@ resource "random_string" "this" {
   special = false
 }
 
-module "standard" {
+module "policy" {
   source = "../../"
 
   name                = "tftest${random_string.this.result}"
@@ -56,7 +56,7 @@ data "aws_iam_policy_document" "s3" {
     ]
 
     resources = [
-      "${module.standard.arn}",
+      "${module.policy.arn}",
     ]
   }
 }