+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
+## 0.3.0
+
+* Add policy document ouputs.
+* Add lifecycle deletion for object lock bucket.
+
 ## 0.2.0
 
-* Supporting versioning
-* Supporing object lock
+* Supporting versioning.
+* Supporing object lock.